# docs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: 筑享开发团队
  description: 筑享移动端和管理端后端接口。除明确标注为公开接口外，受保护接口需在 Authorization 请求头中携带 Bearer 访问令牌。
  license:
    name: 内部使用
  title: 筑享租房平台接口文档
  version: v1
servers:
- description: Generated server url
  url: http://localhost:18080/api
...
paths:
...
  /messages:
    get:
      description: 按消息分类和已读状态筛选当前用户的站内消息。
      operationId: list
      parameters:
      - description: 消息分类：system、lease、lock、bill、appointment 或 repair
        in: query
        name: category
        required: false
        schema:
          type: string
      - description: 是否已读；不传则查询全部
        in: query
        name: isRead
        required: false
        schema:
          type: boolean
      - description: 页码，从 1 开始
        example: 1
        in: query
        name: page
        required: false
        schema:
          type: integer
          format: int64
          default: 1
          minimum: 1
      - description: 每页条数，范围 1-100
        example: 20
        in: query
        name: pageSize
        required: false
        schema:
          type: integer
          format: int64
          default: 20
          maximum: 100
          minimum: 1
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePageDataMessageView"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 分页查询消息
      tags:
      - 消息
  /messages/read:
    delete:
      description: 删除当前用户的全部已读消息，未读消息不受影响。
      operationId: clearRead
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBoolean"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 清空已读消息
      tags:
      - 消息
  /messages/read-all:
    put:
      description: 将当前用户的全部消息标记为已读。
      operationId: markAllRead
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBoolean"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 全部标记已读
      tags:
      - 消息
  /messages/unread-counts:
    get:
      description: 返回全部未读数以及各消息分类的未读数。
      operationId: unreadCounts
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUnreadCounts"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 获取未读消息数
      tags:
      - 消息
  /messages/{id}:
    delete:
      description: 删除当前用户的指定消息。
      operationId: delete
      parameters:
      - description: 消息 ID
        example: message_001
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBoolean"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 删除消息
      tags:
      - 消息
  /messages/{id}/read:
    put:
      description: 将当前用户的指定消息标记为已读。
      operationId: markRead
      parameters:
      - description: 消息 ID
        example: message_001
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBoolean"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 标记单条消息已读
      tags:
      - 消息
...
components:
  schemas:
...
    ApiResponseBoolean:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          type: boolean
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ApiResponsePageDataMessageView:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/PageDataMessageView"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ApiResponseUnreadCounts:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/UnreadCounts"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ApiResponseVoid:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    MessageView:
      type: object
      description: 站内消息
      properties:
        actionTarget:
          type: string
          description: 点击操作目标
        actionType:
          type: string
          description: 点击操作类型
        category:
          type: string
          description: 分类标识
        content:
          type: string
          description: 消息正文
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        iconKey:
          type: string
          description: 前端图标键
        id:
          type: string
          description: 记录 ID
        isRead:
          type: boolean
          description: 是否已读
        title:
          type: string
          description: 标题
...
    PageDataMessageView:
      type: object
      description: 分页数据
      properties:
        hasMore:
          type: boolean
          description: 是否还有下一页
          example: true
        items:
          type: array
          description: 当前页数据
          items:
            $ref: "#/components/schemas/MessageView"
        page:
          type: integer
          format: int64
          description: 当前页码，从 1 开始
          example: 1
        pageSize:
          type: integer
          format: int64
          description: 每页条数
          example: 20
        total:
          type: integer
          format: int64
          description: 总记录数
          example: 42
...
    UnreadCounts:
      type: object
      description: 各分类未读消息数量
      properties:
        appointment:
          type: integer
          format: int64
          description: 预约类未读消息数
        bill:
          type: integer
          format: int64
          description: 账单类未读消息数
        lease:
          type: integer
          format: int64
          description: 租约类未读消息数
        lock:
          type: integer
          format: int64
          description: 门锁类未读消息数
        repair:
          type: integer
          format: int64
          description: 报修类未读消息数
        system:
          type: integer
          format: int64
          description: 系统类未读消息数
        total:
          type: integer
          format: int64
          description: 总记录数或合计数量
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: "访问令牌，格式：Bearer {accessToken}"
      name: Authorization
      scheme: bearer
      type: http